# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Influent API",
    "version": "1.0.0",
    "description": "OpenAPI specification for the Influent backend API (partial)."
  },
  "basePath": "/api",
  "servers": [
    {
      "url": "/api",
      "description": "Base path for the API"
    }
  ],
  "paths": {
...
    "/api/uploads/{filename}": {
      "get": {
        "summary": "Serve uploaded image",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Image file (binary)",
            "content": {
              "image/*": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          }
        }
      }
    },
    "/v1/upload/campaign/{id}/banner": {
      "post": {
        "summary": "Upload campaign banner",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "banner_image": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Uploaded" } },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/v1/upload/campaign/{id}/references": {
      "post": {
        "summary": "Upload campaign references",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "reference_images": {
                    "type": "array",
                    "items": { "type": "string", "format": "binary" }
                  }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Uploaded" } },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/v1/upload/user/{id}/profile": {
      "post": {
        "summary": "Upload user profile image",
        "tags": ["Uploads"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "profile_image": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile_image": { "type": "string" }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}